Fixes Slow Listing Retrieval
Delays the timeout for retrieving to 10ms. In the vast majority of
cases, it will be retrieved in under 10ms. If it still takes longer,
`retrieveListingInfoFromPage` rejects the promise. If the promise is
rejected for a single listing, the listing text is updated (to prevent
an infinite loop if they're on a new page). If the promise is rejected
for 'Get all floats', it tries again.