Fallback Offer Annotate If Seller Send Trade Offer Outside of Extension
As long as they send the correct asset and have the extension running on a device, we will annotate the offer ID regardless of whether the extension was used to create the original offer.

Previously, we'd still verify the asset, but we wouldn't recognize that an open trade offer was active and therefore wouldn't adjust timers.